# apps/ai-gateway/docs/swagger.json
{
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "swagger": "2.0",
  "info": {
    "description": "Multi-provider AI Gateway. Proxies requests to OpenAI, Anthropic, GitHub Copilot, Ollama, Bedrock.",
    "title": "AI Gateway API",
    "contact": {
      "name": "Maintainer",
      "email": "your@email.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0"
  },
  "basePath": "/",
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "description": "Entry point for the AI Gateway's chat completion interface (OpenAI compatible).",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["completions"],
        "summary": "Chat completions",
        "parameters": [
          {
            "description": "Chat Request",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/domain.ChatRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/domain.ChatResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "405": {
            "description": "method not allowed",
            "schema": {
              "type": "string"
            }
          },
          "502": {
            "description": "Bad Gateway",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/v1/embeddings": {
      "post": {
        "description": "Vector generation endpoint.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["embeddings"],
        "summary": "Embeddings",
        "parameters": [
          {
            "description": "Embedding Request",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/domain.EmbeddingRequest"
            }
          },
          {
            "type": "string",
            "description": "Provider name override",
            "name": "X-AI-Provider",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/domain.EmbeddingResponse"
            }
          }
        }
      }
    },
...
    "/v1/models/{provider}": {
      "get": {
        "description": "List available capabilities/models for a provider.",
        "produces": ["application/json"],
        "tags": ["models"],
        "summary": "List models",
        "parameters": [
          {
            "type": "string",
            "description": "Provider name (e.g., openai, anthropic)",
            "name": "provider",
            "in": "path"
          },
          {
            "type": "string",
            "description": "Provider name if not in path",
            "name": "X-AI-Provider",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "domain.ChatDelta": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "tool_calls": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/domain.ToolCallChunk"
          }
        }
      }
    },
    "domain.ChatRequest": {
      "type": "object",
      "properties": {
        "max_tokens": {
          "type": "integer"
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/domain.Message"
          }
        },
        "model": {
          "type": "string"
        },
        "n": {
          "type": "integer"
        },
        "stop": {},
        "stream": {
          "type": "boolean"
        },
        "stream_options": {
          "$ref": "#/definitions/domain.StreamOptions"
        },
        "temperature": {
          "type": "number"
        },
        "tool_choice": {},
        "tools": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/domain.Tool"
          }
        },
        "top_p": {
          "type": "number"
        }
      }
    },
    "domain.ChatResponse": {
      "type": "object",
      "properties": {
        "choices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/domain.Choice"
          }
        },
        "created": {
          "type": "integer"
        },
        "id": {
          "type": "string"
        },
        "model": {
          "type": "string"
        },
        "object": {
          "type": "string"
        },
        "usage": {
          "$ref": "#/definitions/domain.Usage"
        }
      }
    },
    "domain.Choice": {
      "type": "object",
      "properties": {
        "delta": {
          "$ref": "#/definitions/domain.ChatDelta"
        },
        "finish_reason": {
          "type": "string"
        },
        "index": {
          "type": "integer"
        },
        "message": {
          "$ref": "#/definitions/domain.Message"
        }
      }
    },
    "domain.EmbeddingData": {
      "type": "object",
      "properties": {
        "embedding": {
          "type": "array",
          "items": {
            "type": "number"
          }
        },
        "index": {
          "type": "integer"
        },
        "object": {
          "type": "string"
        }
      }
    },
    "domain.EmbeddingRequest": {
      "type": "object",
      "properties": {
        "input": {
          "description": "Can be string or []string"
        },
        "model": {
          "type": "string",
          "example": "text-embedding-3-small"
        }
      }
    },
    "domain.EmbeddingResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/domain.EmbeddingData"
          }
        },
        "model": {
          "type": "string"
        },
        "object": {
          "type": "string"
        },
        "usage": {
          "$ref": "#/definitions/domain.Usage"
        }
      }
    },
    "domain.FunctionCall": {
      "type": "object",
      "properties": {
        "arguments": {
          "description": "JSON string",
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "domain.FunctionChunk": {
      "type": "object",
      "properties": {
        "arguments": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "domain.FunctionDefinition": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "parameters": {
          "description": "JSON Schema"
        }
      }
    },
    "domain.Message": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "role": {
          "description": "\"system\" | \"user\" | \"assistant\" | \"tool\"",
          "type": "string"
        },
        "tool_call_id": {
          "type": "string"
        },
        "tool_calls": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/domain.ToolCall"
          }
        }
      }
    },
...
    "domain.StreamOptions": {
      "type": "object",
      "properties": {
        "include_usage": {
          "type": "boolean"
        }
      }
    },
...
    "domain.Tool": {
      "type": "object",
      "properties": {
        "function": {
          "$ref": "#/definitions/domain.FunctionDefinition"
        },
        "type": {
          "description": "always \"function\"",
          "type": "string"
        }
      }
    },
    "domain.ToolCall": {
      "type": "object",
      "properties": {
        "function": {
          "$ref": "#/definitions/domain.FunctionCall"
        },
        "id": {
          "type": "string"
        },
        "type": {
          "description": "always \"function\"",
          "type": "string"
        }
      }
    },
    "domain.ToolCallChunk": {
      "type": "object",
      "properties": {
        "function": {
          "$ref": "#/definitions/domain.FunctionChunk"
        },
        "id": {
          "type": "string"
        },
        "index": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "domain.Usage": {
      "type": "object",
      "properties": {
        "completion_tokens": {
          "type": "integer"
        },
        "prompt_tokens": {
          "type": "integer"
        },
        "total_tokens": {
          "type": "integer"
        }
      }
    },
...
  }
}